+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   function result via an argument. Used when
   fmt.F_string_result_as_arg is defined.  Eliminated
   meta["is_result"].
+- Remove stmt.c_result_var. Remove custom code in wrapf and replaced
+  with data in FStmts.
 
 ## v0.13.0 - 2023-08-23
 ### Added